This is a program that for a given history identifies whether this history is Serializable.
Specifically, it uses Conflict graph to check if the history belongs to CSR.
The input format is w 1 x r 2 x w 2 y, i.e. a totally ordered history. 
All transaction numbers start at 1 and are consecutive. The commit operations are left out for simplicity, but all transactions do commit. The output (true/false) is written as an output.